separate intersection module + service
socket communication will established only, if the module is loaded
