+++ web-app/src/app/app.effects.ts
@@ -1,15 +1,7 @@
 import { CoreEffects } from './core/store/core.effects';
 import { DataTypesEffects } from './core/datatypes/store/data-types.effects';
 import { GenericDataEffects } from './eep/generic-data/store/generic-data.effects';
-import { IntersectionEffects } from './eep/intersection/store/intersection.effects';
 import { SignalEffects } from './eep/signals/store/signal.effects';
 import { TrainEffects } from './eep/trains/store/train.effects';
 
-export const effects = [
-  CoreEffects,
-  DataTypesEffects,
-  GenericDataEffects,
-  IntersectionEffects,
-  SignalEffects,
-  TrainEffects,
-];
+export const effects = [CoreEffects, DataTypesEffects, GenericDataEffects, SignalEffects, TrainEffects];
