+++ web-app/src/app/app.reducers.ts
@@ -3,7 +3,6 @@ import { ActionReducerMap } from '@ngrx/store';
 import * as fromCore from './core/store/core.reducers';
 import * as fromDataTypes from './core/datatypes/store/data-types.reducers';
 import * as fromGenericData from './eep/generic-data/store/generic-data.reducers';
-import * as fromIntersection from './eep/intersection/store/intersection.reducers';
 import * as fromSignal from './eep/signals/store/signal.reducers';
 import * as fromTrain from './eep/trains/store/train.reducer';
 import * as fromStatistics from './eep/statistics/store/statistics.reducer';
@@ -12,7 +11,6 @@ export interface State {
   core: fromCore.State;
   dataTypes: fromDataTypes.State;
   genericData: fromGenericData.State;
-  intersection: fromIntersection.State;
   signal: fromSignal.State;
   train: fromTrain.State;
   statistics: fromStatistics.State;
@@ -22,7 +20,6 @@ export const reducers: ActionReducerMap<State> = {
   core: fromCore.reducer,
   dataTypes: fromDataTypes.reducer,
   genericData: fromGenericData.reducer,
-  intersection: fromIntersection.reducer,
   signal: fromSignal.reducer,
   train: fromTrain.reducer,
   statistics: fromStatistics.reducer,
