+++ web-app/src/app/client/modules/intersection/intersection-detail/intersection.component.html
