+++ web-app/src/app/client/modules/intersection/intersection-list/intersections.component.css
