+++ web-app/src/app/client/modules/intersection/intersection-list/intersections.component.ts
@@ -3,13 +3,13 @@ import { Observable } from 'rxjs';
 import { select, Store } from '@ngrx/store';
 
 import { Intersection } from '../models/intersection.model';
-import * as fromRoot from '../../../app.reducers';
-import * as fromIntersection from '../../intersection/store/intersection.reducers';
+import * as fromRoot from '../../../../app.reducers';
+import * as fromIntersection from '../store/intersection.reducers';
 import { Router } from '@angular/router';
 import { IntersectionHelper } from '../intersection-helper';
 import { MatDialog } from '@angular/material/dialog';
-import { LuaSettings } from '../../../shared/model/lua-settings';
-import { LuaSettingChangeEvent } from '../../../shared/model/lua-setting-change-event';
+import { LuaSettings } from '../../../../shared/model/lua-settings';
+import { LuaSettingChangeEvent } from '../../../../shared/model/lua-setting-change-event';
 import * as IntersectionAction from '../store/intersection.actions';
 
 @Component({
