+++ web-app/src/app/client/modules/intersection/intersection-routing.module.ts
@@ -5,17 +5,12 @@ import { IntersectionsComponent } from './intersection-list/intersections.compon
 import { IntersectionComponent } from './intersection-detail/intersection.component';
 
 const intersectionRoutes: Routes = [
-  {path: '', component: IntersectionsComponent, pathMatch: 'full'},
-  {path: ':id', component: IntersectionComponent, pathMatch: 'full'},
+  { path: '', component: IntersectionsComponent, pathMatch: 'full' },
+  { path: ':id', component: IntersectionComponent, pathMatch: 'full' },
 ];
 
 @NgModule({
-  imports: [
-    RouterModule.forChild(intersectionRoutes)
-  ],
-  exports: [
-    RouterModule
-  ]
+  imports: [RouterModule.forChild(intersectionRoutes)],
+  exports: [RouterModule],
 })
-export class IntersectionRoutingModule {
-}
+export class IntersectionRoutingModule {}
