+++ web-app/src/app/client/modules/intersection/intersection-switching/intersection-switching.component.html
