+++ web-app/src/app/client/modules/intersection/lane-overview/lane-overview.component.html
