+++ web-app/src/app/client/modules/intersection/lane-queue/lane-queue.component.html
