+++ web-app/src/app/client/modules/intersection/models/count-type.enum.ts
