+++ web-app/src/app/client/modules/intersection/models/direction.model.ts
