+++ web-app/src/app/client/modules/intersection/models/intersection-lane.model.ts
