+++ web-app/src/app/client/modules/intersection/models/intersection-switching.model.ts
