+++ web-app/src/app/client/modules/intersection/models/intersection-traffic-light.model.ts
