+++ web-app/src/app/client/modules/intersection/models/intersection.model.ts
