+++ web-app/src/app/client/modules/intersection/models/phase.enum.ts
