+++ web-app/src/app/client/modules/intersection/models/traffic-type.enum.ts
